Refactor CLA files resolution to the commands module
